full crud with authorization

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 import asyncHandler from 'express-async-handler';
 import Goal from '../models/goalModel.js';
+import User from '../models/userModel.js';
 //@desc get all goals
 //@route GET /api/goals
 //@access Public
 const getGoals = asyncHandler(async(req, res) => {
 
-  const goals = await Goal.find({});
+  const goals = await Goal.find({ user: req.user._id });
     res.status(200).json(goals);
 
   })
@@ -19,7 +20,9 @@ const setGoal = asyncHandler( async(req, res) => {
        throw new Error('Please enter a goal');
     }
     const goal = await Goal.create({
-      text: req.body.text});
+      text: req.body.text,
+      user: req.user._id
+    });
      res.status(201).json(goal);
   })
 
@@ -31,6 +34,17 @@ const updateGoal = asyncHandler(async(req, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found');
+    }
+    const user = await User.findById(req.user._id);
+    // check for valid user
+    if(!user){
+        res.status(400)
+        throw new Error('User not found');
+    }
+    // check if goal belongs to user
+    if(goal.user.toString() !== user._id.toString()){
+        res.status(401)
+        throw new Error('Not authorized');
     }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
@@ -46,6 +60,17 @@ const deleteGoal = asyncHandler(async(req, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found');
+    }
+    const user = await User.findById(req.user._id);
+    // check for valid user
+    if(!user){
+        res.status(400)
+        throw new Error('User not found');
+    }
+    // check if goal belongs to user
+    if(goal.user.toString() !== user._id.toString()){
+        res.status(401)
+        throw new Error('Not authorized');
     }
     const deleteGoal = await Goal.findByIdAndDelete(req.params.id);
     res.status(200).json({id: req.params.id});
